+++ backend/canisters/user_index/impl/src/lifecycle/post_upgrade.rs
@@ -23,7 +23,6 @@ fn post_upgrade(args: Args) {
     canister_logger::init_with_logs(data.test_mode, errors, logs, traces);
 
     let env = init_env(data.rng_seed, data.oc_key_pair.is_initialised());
-
     init_cycles_dispenser_client(data.cycles_dispenser_canister_id, data.test_mode);
     init_state(env, data, args.wasm_version);